of_folder/requirements.txt
# Core LangChain/LangGraph dependencies (Latest versions)
langchain==0.3.2
langgraph==0.2.28
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.5
langchain-text-splitters==0.3.0
langsmith==0.1.17

# Machine Learning & LoRA (Updated versions)
transformers==4.45.2
peft==0.13.2
torch==2.4.1
tokenizers==0.20.0
datasets==2.21.0
accelerate==0.34.2
huggingface-hub==0.25.1

# Social Media APIs (Updated versions)
tweepy==4.14.0
praw==7.7.1
instagrapi==2.0.0
requests-oauthlib==1.3.1
openai==1.51.2
anthropic==0.34.2
replicate==0.25.1

# Image Processing & Computer Vision
opencv-python==4.9.0.80
Pillow==10.1.0

# MCP (Model Context Protocol)
mcp==1.0.0

# Additional LangChain integrations
langchain-anthropic==0.1.23
langchain-experimental==0.0.64
langchain-google-genai==1.0.10

# Data & Analytics
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Database & Storage
sqlalchemy==2.0.25
alembic==1.13.1
# PostgreSQL + Azure/Vertex Integration
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
pgvector==0.2.4

# Vector Database & Retrieval
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Web & API
fastapi==0.108.0
uvicorn==0.25.0
httpx==0.26.0
aiohttp==3.9.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1

# Utilities
rich==13.7.0
typer==0.9.0
schedule==1.2.1
pytz==2023.3.post1

# Testing & Monitoring
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0
pytest-cov==4.1.0
prometheus-client==0.20.0
jaeger-client==4.8.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
psycopg2-binary==2.9.9

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Windows-specific
pywin32==306; sys_platform == "win32" 